intel test working
Conflicts:
	opal/datatype/cuda/opal_datatype_unpack_cuda_wrapper.cu
	opal/mca/btl/smcuda/btl_smcuda.c

fix a bug when buffer is not big enough for whole ddt

Conflicts:
	opal/datatype/cuda/opal_datatype_unpack_cuda_wrapper.cu

if data in different gpu, instead of copy direct from one to the other,
we do a D2D copy

Conflicts:
	opal/datatype/cuda/opal_datatype_unpack_cuda_wrapper.cu
	test/datatype/Makefile.am

now we can use cudamemcpy2d

Conflicts:
	opal/datatype/cuda/opal_datatype_unpack_cuda_wrapper.cu

enable zero copy + fix GPU buffer bug

Conflicts:
	opal/datatype/cuda/opal_datatype_unpack_cuda_wrapper.cu

put pipeline size into mca

+++ opal/datatype/cuda/opal_datatype_unpack_cuda_kernel.cu
@@ -277,9 +277,9 @@ __global__ void opal_generic_simple_unpack_cuda_iov_kernel( ddt_cuda_iov_dist_t*
             _destination_tmp = dst + threadIdx.x * alignment;
 #if !defined (OPAL_DATATYPE_CUDA_DRY_RUN)
                 if (alignment == ALIGNMENT_DOUBLE) {
-                    *((double *)_destination_tmp) = *((double *)_source_tmp);
+                    *((long *)_destination_tmp) = *((long *)_source_tmp);
                 } else if (alignment == ALIGNMENT_FLOAT) {
-                    *((float *)_destination_tmp) = *((float *)_source_tmp);
+                    *((int *)_destination_tmp) = *((int *)_source_tmp);
                 } else {
                     * _destination_tmp = *_source_tmp;
                 }
